"commit 03/12"

+++ src/main/java/com/example/testingsp/Repository/FileEntityRepo.java
@@ -3,5 +3,5 @@ package com.example.testingsp.Repository;
 import com.example.testingsp.DTO.FileEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FileEntityRepo extends JpaRepository<FileEntity,Integer> {
+public interface FileEntityRepo extends JpaRepository<FileEntity,Long> {
 }
